Cambios en programacion de vistas

+++ Models/ViewModels/UnidadAdministrativasViewModel.cs
@@ -8,7 +8,7 @@ namespace ProyectoPOA.Models.ViewModels
     public class UnidadAdministrativasViewModel
     {
         public int Id { get; set; }
-        public int Clave { get; set; }
+        public short Clave { get; set; }
         public string Nombre { get; set; }
         public string NombreEncargado { get; set; }
         public string NombreSuperior { get; set; }
